feat: create DeleteModal

+++ src/app/api/projects/[id]/route.ts
@@ -3,6 +3,11 @@ const prisma = new PrismaClient();
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
   const id = params.id;
+  const secretKey = req.headers.get("x-api-key");
+
+  if (!secretKey || secretKey !== process.env.NEXT_PUBLIC_API_KEY) {
+    throw new Error("Sem autorização");
+  }
 
   try {
     const project = await prisma.project.findMany({
@@ -19,6 +24,11 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
 
 export async function DELETE(req: Request, { params }: { params: { id: string } }) {
   const id = params.id;
+  const secretKey = req.headers.get("x-api-key");
+
+  if (!secretKey || secretKey !== process.env.NEXT_PUBLIC_API_KEY) {
+    throw new Error("Sem autorização");
+  }
 
   try {
     await prisma.project.delete({ where: { id: id } });
